apps/aneca-advisor/backend/requirements.txt
flask
flask-cors
gunicorn
requests
psycopg2-binary
pytest>=8.0
-e ../../../libs/apps-platform-sdk
